Option and Discount fixes

+++ Presentation/Controllers/CartsController.cs
@@ -1,6 +1,9 @@
-﻿using Contracts.Model.Product;
+﻿using System.Net;
+using System.Security.Claims;
+using Contracts.Model.Product;
 using Contracts.Model.Product.Response;
 using Contracts.Model.Product.UserInteract;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
 using Services.Abstractions;
@@ -9,6 +12,7 @@ namespace Presentation.Controllers;
 
 [ApiController]
 [Route("api/[controller]")]
+[Authorize]
 public class CartsController : ControllerBase {
     private readonly IServiceManager _serviceManager;
 
@@ -17,43 +21,79 @@ public class CartsController : ControllerBase {
     }
 
     [HttpGet]
+    [Authorize(Roles = "Admin")]
     public async Task<IActionResult> GetCarts(CancellationToken cancellationToken) {
         IEnumerable<CartResponse> carts = await _serviceManager.CartService.GetAllAsync(cancellationToken);
         return Ok(carts);
     }
+    
+    [HttpGet("product/{productCode}/option/{optionId}/discountPreviews")]
+    public async Task<IActionResult> GetDiscountPreviews(string productCode, int? optionId, CancellationToken cancellationToken)
+    {
+        string? userId = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+        if (userId == null)
+            return StatusCode((int) HttpStatusCode.Unauthorized);
+        IEnumerable<ProductsDiscountResponse> discountPreviews = await _serviceManager.CartService.GetDiscountsPreviewAsync(userId, productCode, optionId, cancellationToken);
+        return Ok(discountPreviews);
+    }
 
     [HttpGet("{userId}")]
+    [Authorize(Roles = "Admin")]
     public async Task<IActionResult> GetCart(string userId, CancellationToken cancellationToken) {
         CartResponse cart = await _serviceManager.CartService.GetByIdAsync(userId, cancellationToken);
         return Ok(cart);
     }
     
-    [HttpPut("{userId}/AddProductToCart")]
-    public async Task<IActionResult> AddProductToCartAsync(string userId, [FromBody] PurchaseDetailDto cartItem, CancellationToken cancellationToken) {
-        CartResponse cart = await _serviceManager.CartService.AddProductToCartAsync(userId, cartItem, cancellationToken);
+    [HttpGet("myCart")]
+    public async Task<IActionResult> GetMyCart(CancellationToken cancellationToken) {
+        string? userId = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+        if (userId == null)
+            return StatusCode((int) HttpStatusCode.Unauthorized);
+        CartResponse cart = await _serviceManager.CartService.GetByIdAsync(userId,cancellationToken);
+        return Ok(cart);
+    }
+    
+    [HttpPut("discount/{discountCode}/AddProductToCart")]
+    public async Task<IActionResult> AddProductToCartAsync([FromBody] PurchaseDetailDto cartItem, string? discountCode, CancellationToken cancellationToken) {
+        string? userId = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+        if (userId == null)
+            return StatusCode((int) HttpStatusCode.Unauthorized);
+        CartResponse cart = await _serviceManager.CartService.AddProductToCartAsync(userId, discountCode, cartItem, cancellationToken);
         return Ok(cart);
     }
 
-    [HttpPut("{userId}/AddPackToCart")]
-    public async Task<IActionResult> AddPackToCartAsync(string userId, [FromBody] List<PurchaseDetailDto> cartItemsItems, CancellationToken cancellationToken) {
+    [HttpPut("AddPackToCart")]
+    public async Task<IActionResult> AddPackToCartAsync([FromBody] List<PurchaseDetailDto> cartItemsItems, CancellationToken cancellationToken) {
+        string? userId = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+        if (userId == null)
+            return StatusCode((int) HttpStatusCode.Unauthorized);
         CartResponse cart = await _serviceManager.CartService.AddPackToCartAsync(userId, cartItemsItems, cancellationToken);
         return Ok(cart);
     }
     
-    [HttpPut("{userId}/product/{productId}/adjust")]
-    public async Task<IActionResult> AdjustCartItemQuantityAsync(string userId, string productId, int adjust, CancellationToken cancellationToken) {
+    [HttpPut("product/{productId}/adjust")]
+    public async Task<IActionResult> AdjustCartItemQuantityAsync(string productId, int adjust, CancellationToken cancellationToken) {
+        string? userId = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+        if (userId == null)
+            return StatusCode((int) HttpStatusCode.Unauthorized);
         CartResponse cart = await _serviceManager.CartService.AdjustCartItemQuantityAsync(userId, productId, adjust, cancellationToken);
         return Ok(cart);
     }
     
-    [HttpPut("{userId}/product/{productId}/remove")]
-    public async Task<IActionResult> RemoveCartItemAsync(string userId, string productId, CancellationToken cancellationToken) {
+    [HttpPut("product/{productId}/remove")]
+    public async Task<IActionResult> RemoveCartItemAsync(string productId, CancellationToken cancellationToken) {
+        string? userId = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+        if (userId == null)
+            return StatusCode((int) HttpStatusCode.Unauthorized);
         CartResponse cart = await _serviceManager.CartService.RemoveCartItemAsync(userId, productId, cancellationToken);
         return Ok(cart);
     }
     
-    [HttpPut("{userId}/clear")]
-    public async Task<IActionResult> ClearCartAsync(string userId, CancellationToken cancellationToken) {
+    [HttpPut("clear")]
+    public async Task<IActionResult> ClearCartAsync(CancellationToken cancellationToken) {
+        string? userId = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+        if (userId == null)
+            return StatusCode((int) HttpStatusCode.Unauthorized);
         CartResponse cart = await _serviceManager.CartService.ClearCartAsync(userId, cancellationToken);
         return Ok(cart);
     }
